+++ lib/tensorflow.rb
@@ -9,12 +9,18 @@ require "net/http"
 require "tempfile"
 
 # modules
-require "tensorflow/utils"
+require "tensorflow/audio"
+require "tensorflow/bitwise"
 require "tensorflow/context"
+require "tensorflow/image"
+require "tensorflow/io"
+require "tensorflow/linalg"
 require "tensorflow/math"
 require "tensorflow/ops"
 require "tensorflow/raw_ops"
+require "tensorflow/strings"
 require "tensorflow/tensor"
+require "tensorflow/utils"
 require "tensorflow/variable"
 require "tensorflow/version"
 
@@ -41,6 +47,7 @@ module TensorFlow
     include Utils
 
     extend Forwardable
+    def_delegators Linalg, :matmul
     def_delegators Math, :abs, :acos, :acosh, :add, :add_n, :argmax, :argmin, :asin, :asinh, :atan, :atan2, :atanh, :cos, :cosh, :cumsum, :divide, :equal, :exp, :floor, :greater, :greater_equal, :less, :less_equal, :logical_and, :logical_not, :logical_or, :maximum, :minimum, :multiply, :negative, :not_equal, :pow, :reduce_all, :reduce_any, :reduce_logsumexp, :reduce_max, :reduce_mean, :reduce_min, :reduce_prod, :reduce_sum, :round, :scalar_mul, :sigmoid, :sign, :sin, :sinh, :sqrt, :square, :subtract, :tan, :tanh, :truediv
 
     def library_version